Fee calculated with int internally and converted to double to fit API

+++ ParkingAPI/Services/IFeeCalculator.cs
@@ -4,6 +4,6 @@ namespace ParkingAPI.Services
 {
     public interface IFeeCalculator
     {
-        public double CalculateFee(SpaceOccupancy occupancyPeriod);
+        public int CalculateFee(SpaceOccupancy occupancyPeriod);
     }
 }
